+++ CotEditor/Sources/SyntaxMappingConflictView.swift
@@ -124,7 +124,7 @@ private struct ConflictTable: View {
                 TableColumn(String(localized: "Used syntax", table: "SyntaxMappingConflict", comment: "table column header"), value: \.primarySyntax) {
                     Text($0.primarySyntax).fontWeight(.semibold)
                 }
-                TableColumn(String(localized: "Duplicated syntaxes", table: "SyntaxMappingConflict", comment: "table column header")) {
+                TableColumn(String(localized: "Duplicated syntaxes", table: "SyntaxMappingConflict", comment: "table column header"), sortUsing: KeyPathComparator(\.duplicatedSyntaxes.first)) {
                     Text($0.duplicatedSyntaxes, format: .list(type: .and, width: .narrow))
                 }
             }